Tier B Phase 2 (Slice 1): move framework-free widget core to @mcpjam/sdk
Relocate two framework-free, zero-inspector-dep modules into a new
@mcpjam/sdk/widget-runtime subpath, leaving back-compat re-export shims at the
original inspector paths so consumers are unchanged:

- tool-visibility (SEP-1865 getToolVisibility / isVisibleToModelOnly /
  isVisibleToAppOnly) — zero imports.
- mcp-apps-logging-transport (LoggingTransport) — type-only @modelcontextprotocol/
  client imports, already an SDK dep.

SDK wiring: new src/widget-runtime/{tool-visibility,logging-transport,index}.ts,
tsup entry, and a ./widget-runtime export. The inspector resolves it from the
built dist (every CI gate builds the SDK before consuming it: root `test`,
lint.yml `typecheck`, and inspector `build` via build:sdk).

This is the smallest, safest Phase 2 slice — it establishes the
@mcpjam/sdk/widget-runtime pattern end-to-end (iframe-sandbox-policy and
host-app-bridge follow). Pure relocation behind stable import paths.

Verification: SDK build + typecheck clean; inspector typecheck:client + renderer
tier-b guard pass; 134 tests pass (host-app-bridge 26, sandboxed-iframe 22,
mcp-apps-renderer 86).
